+++ doc/services/portability/posix/kconfig/index.rst
@@ -9,7 +9,6 @@ implementation of the POSIX API.
 * :kconfig:option:`CONFIG_DYNAMIC_THREAD`
 * :kconfig:option:`CONFIG_DYNAMIC_THREAD_POOL_SIZE`
 * :kconfig:option:`CONFIG_EVENTFD`
-* :kconfig:option:`CONFIG_EVENTFD_MAX`
 * :kconfig:option:`CONFIG_FDTABLE`
 * :kconfig:option:`CONFIG_GETOPT_LONG`
 * :kconfig:option:`CONFIG_MAX_PTHREAD_SPINLOCK_COUNT`
@@ -38,3 +37,4 @@ implementation of the POSIX API.
 * :kconfig:option:`CONFIG_POSIX_SEM_VALUE_MAX`
 * :kconfig:option:`CONFIG_TIMER_CREATE_WAIT`
 * :kconfig:option:`CONFIG_THREAD_STACK_INFO`
+* :kconfig:option:`CONFIG_ZVFS_EVENTFD_MAX`
