+++ lib/posix/options/Kconfig.compat
@@ -5,26 +5,15 @@
 
 menu "Miscellaneous POSIX-related options"
 
-menuconfig EVENTFD
+config EVENTFD
 	bool "Support for eventfd"
 	depends on !NATIVE_APPLICATION
-	select POLL
-	select FDTABLE
+	select ZVFS
+	select ZVFS_EVENTFD
 	default y if POSIX_API
 	help
 	  Enable support for event file descriptors, eventfd. An eventfd can
 	  be used as an event wait/notify mechanism together with POSIX calls
 	  like read, write and poll.
-
-if EVENTFD
-
-config EVENTFD_MAX
-	int "Maximum number of eventfd's"
-	default 1
-	range 1 4096
-	help
-	  The maximum number of supported event file descriptors.
-
-endif # EVENTFD
 
 endmenu
