+++ tests/net/lib/http_server/crime/prj.conf
@@ -11,7 +11,7 @@ CONFIG_ZTEST_STACK_SIZE=1024
 
 CONFIG_ZVFS_OPEN_MAX=10
 CONFIG_REQUIRES_FULL_LIBC=y
-CONFIG_EVENTFD_MAX=10
+CONFIG_ZVFS_EVENTFD_MAX=10
 CONFIG_NET_MAX_CONTEXTS=10
 CONFIG_NET_MAX_CONN=10
