+++ tests/net/lib/http_server/prototype/prj.conf
@@ -9,7 +9,7 @@ CONFIG_ENTROPY_GENERATOR=y
 CONFIG_TEST_RANDOM_GENERATOR=y
 CONFIG_ZVFS_OPEN_MAX=10
 CONFIG_REQUIRES_FULL_LIBC=y
-CONFIG_EVENTFD_MAX=10
+CONFIG_ZVFS_EVENTFD_MAX=10
 CONFIG_NET_MAX_CONTEXTS=10
 CONFIG_NET_MAX_CONN=10
