+++ tests/net/lib/http_server/tls/prj.conf
@@ -41,7 +41,7 @@ CONFIG_ZVFS_OPEN_MAX=32
 CONFIG_NET_SOCKETS_POLL_MAX=32
 CONFIG_ZVFS_OPEN_MAX=32
 CONFIG_REQUIRES_FULL_LIBC=y
-CONFIG_EVENTFD_MAX=10
+CONFIG_ZVFS_EVENTFD_MAX=10
 CONFIG_NET_MAX_CONTEXTS=10
 CONFIG_NET_MAX_CONN=10
